move "...-utils" modules to "utils" sub-dir
