+++ lib/auth.js
@@ -11,8 +11,8 @@ const mkdirp = pify(require('mkdirp'));
 
 // local modules
 
-const getOrigin = require('./url-utils').getOrigin;
-const jsonUtils = require('./json-utils');
+const getOrigin = require('./utils/url').getOrigin;
+const jsonUtils = require('./utils/json');
 const values = require('./values');
 
 // this module
