+++ lib/scope.js
@@ -11,8 +11,8 @@ const findUp = require('find-up');
 // local modules
 
 const CONFIG_FILE = require('./values').CONFIG_FILE;
-const isURL = require('./url-utils').isURL;
-const jsonUtils = require('./json-utils');
+const isURL = require('./utils/url').isURL;
+const jsonUtils = require('./utils/json');
 
 // this module
