+++ lib/whoami.js
@@ -6,8 +6,8 @@ const request = require('request');
 
 // local modules
 
-const jsonUtils = require('./json-utils');
-const toHTTPS = require('./url-utils').toHTTPS;
+const jsonUtils = require('./utils/json');
+const toHTTPS = require('./utils/url').toHTTPS;
 const values = require('./values');
 
 // this module
